Users: Correct `get_the_author_posts()` to always return an integer.
This matches the documented `@return` type.

Follow-up to [5638], [13576], [36085], [60296].

Props justlevine.
See #63268.
Built from https://develop.svn.wordpress.org/trunk@60299


git-svn-id: http://core.svn.wordpress.org/trunk@59635 1a063a9b-81f0-0310-95a4-ce76da25c4cd

+++ wp-includes/author-template.php
@@ -286,7 +286,7 @@ function get_the_author_posts() {
 	if ( ! $post ) {
 		return 0;
 	}
-	return count_user_posts( $post->post_author, $post->post_type );
+	return (int) count_user_posts( $post->post_author, $post->post_type );
 }
 
 /**
